Basic HeroBus commands and responses tested working
I had to switch from using the I2C FIFO to regular I2C interrupts, because
the receive FIFO interrupt doesn't seem to be working.  Checking the
registers with a debugger, the interrupt was enabled, and the flag was
set, but the ISR would never get called.  The ISR gets called for fifo
read interrupts (it's the same ISR), so I know that the CPU level enable
is working.  I checked the errata for the part, and nothing was mentioned
with the I2C module, so either I'm doing something wrong with the
configuration of the I2C module, or there is an undocumented errata with
this part that prevents the I2C FIFO receive interrupt from working.
Either way, it works with regular interrupts, and HeroBus traffic should
be relatively light, so the slightly reduced performance of not using the
FIFOs shouldn't really hurt us