shared: move Shared_Boot (from AZ and Boot project) to shared folder
There was a small diff between the files, which is likely a copy&paste error. The diassembled code was the same for those files. The version on AZBootloader was used. The Here is the diff output:
```
diff AZBootloader/Shared_Boot.c Bootloader/Shared_Boot.c
86d85
<           // don't overwrite the memory space I am using.
90c89
<         	  *(Uint16 *)BlockHeader.DestAddr++;
---
>         	  BlockHeader.DestAddr++;
```

Boot.h was also moved as it was identical.

Makefile and include in main where adjusted for the new location