parameters: change all parameters to REAL32 (float) type
This is required because a bug on pymavlink doesn't allow us to handle floating point values in python via the mavlink interface. Look at the following issue for more info:
https://github.com/mavlink/mavlink/issues/404

Since we do not use this parameters in the gimbal code (except for sw_ver), it doesn't affect the gimbal code itself much.

The following changes where made:
* Change UINT32 parameters to REAL32
* Change flash_param_struct to have floats
* Initialization of flash_param struct uses floats now
* update the version number code to save it as a int disguised as a float