flash: add support for migrations of flash parameters
If a struct order has been altered, flash verification will pass but wrong params will be loaded.
This handles different versions of flash parameter structs by doing a forwards migration of the existing values into the positions in the new struct layout.
The migration is one way, as the application relies on the different structures to be compiled in.