gopro_interface: prelim sketch towards simplifying protocol logic
THIS IS AN UNTESTED WIP. JUST FYI. ALL THE BEST.

generally, this is intended to retain current functionality, but reorg in preparation for adding support for multiple gopro protocol versions. 

a variety of the states can be collapsed into a set of procedural routines since all needed info is available to be processed. further, we can wait on in-progress i2c transactions and dispatch completion events in a single point, providing a clearer point to pass off messages to protocol handlers. not introducing the notion of a protocol handler yet.