gopro_interface: observe capture mode on plug-in, get/set 'capture mode' cmds
Gimbal keeps an internal capture mode to determine which I2C command to
send to the GoPro to start/stop recording (for Hero4) and set the
MAVLink heartbeat message to 'recording'. The capture mode state is set
when the GoPro is initially plugged in and when completing 'GET/SET
capture mode' cmds.