gopro: don't send capture mode polling responses, since we report this via heartbeat now
also, make polling for capture mode more general, so we can get entire cam status on hero 3. this gives us capture mode, but also allows us to sniff whether an sd card is likely to be inserted. we'll use this to fix the case in which trigger shutter commands to the hero 3 succeed even when no sd card is inserted.